requirements.txt
pandas
numpy
scikit-learn
bentoml
pydantic
rdkit-pypi
